Changed electron-releases to be a devDependency and generated e2cn.js via npm run importdata to reducce the package size penalty.
